+++ DataLayerService/IDepartmentService.cs
@@ -8,6 +8,6 @@ namespace WebAPI.Data_Layer_Service
         List<Department> GetDepartments();
         void AddDepartment(Department dep);
         void UpdateDepartment(Department dep);
-        void DeleteDepartment(int id);
+        void DeleteDepartment(long id);
     }
 }
